# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface MedicalReport {
  id: number;
  filename: string;
  summary: string;
  diagnosis: string;
  medications: string[];
  followupInstructions: string[];
  warnings: string[];
  language: string;
  createdAt: string;
}

export type MedicalStatsTopLanguagesItem = {
  language: string;
  count: number;
};

export interface MedicalStats {
  totalReports: number;
  recentReports: number;
  topLanguages: MedicalStatsTopLanguagesItem[];
}