# lib/api-zod/src/generated/types/medicalReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MedicalReport {
  id: number;
  filename: string;
  summary: string;
  diagnosis: string;
  medications: string[];
  followupInstructions: string[];
  warnings: string[];
  language: string;
  createdAt: string;
}