# lib/api-zod/src/generated/types/medicalStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MedicalStatsTopLanguagesItem } from "./medicalStatsTopLanguagesItem";

export interface MedicalStats {
  totalReports: number;
  recentReports: number;
  topLanguages: MedicalStatsTopLanguagesItem[];
}